requirements.txt
streamlit
streamlit-folium
folium
opencv-python
pillow
tensorflow
torch
torchvision
ultralytics
scikit-learn
scipy
requests
python-dotenv